+++ pages/index.js
@@ -1,35 +1,16 @@
-import AsanaCard from "../components/AsanaCard";
-import asanas from "./db.json";
-import {useState, useEffect} from "react";
-import Navigation from "../components/Navbar/Navigation";
+import styled from "styled-components";
 
-export default function Home({favorites, setFavorites}) {
-  const [asana, setAsana] = useState(null);
-  const [reload, setReload] = useState(false);
-  function reloader() {
-    setReload(!reload);
-    console.log("fire");
-  }
-  useEffect(() => {
-    function getRandomAsana(array) {
-      const randomIndex = Math.floor(Math.random() * asanas.length);
-      const randomAsana = array[randomIndex];
-      return randomAsana;
-    }
-    setAsana(getRandomAsana(asanas));
-  }, [reload]);
-
-  return (
-    <>
-      {asana && (
-        <AsanaCard
-          asana={asana}
-          favorites={favorites}
-          setFavorites={setFavorites}
-          reloader={reloader}
-        />
-      )}
-      <Navigation />
-    </>
-  );
+export default function Home() {
+  return <StyledMainPage>Test</StyledMainPage>;
 }
+
+const StyledMainPage = styled.div`
+  height: 100vh;
+  width: 100vw;
+  background-color: linear-gradient(
+    45deg,
+    rgba(0, 113, 188, 1) 25%,
+    rgba(63, 143, 213, 1) 35%,
+    rgba(53, 87, 117, 1) 90%
+  );
+`;
